draw: use 32-bit surface index in edge_t

+++ include/r_shared.h
@@ -147,7 +147,7 @@ typedef struct edge_s {
     fixed16_t u;
     fixed16_t u_step;
     struct edge_s *prev, *next;
-    unsigned short surfs[2];
+    unsigned int surfs[2];
     struct edge_s *nextremove;
     float nearzi;
     medge_t *owner;
